--- backend/src/test/scala/com/softwaremill/bootzooka/email/DummyEmailSendingServiceSpec.scala
@@ -1,13 +1,0 @@
-package com.softwaremill.bootzooka.email
-
-import com.softwaremill.bootzooka.email.application.DummyEmailService
-import com.softwaremill.bootzooka.email.domain.EmailContentWithSubject
-import org.scalatest.{FlatSpec, Matchers}
-
-class DummyEmailSendingServiceSpec extends FlatSpec with Matchers {
-  it should "send scheduled email" in {
-    val service = new DummyEmailService
-    service.scheduleEmail("test@sml.com", new EmailContentWithSubject("content", "subject"))
-    service.wasEmailSent("test@sml.com", "subject") should be(true)
-  }
-}
