+++ backend/src/test/scala/com/softwaremill/bootzooka/user/domain/UserSpec.scala
@@ -1,6 +1,13 @@
-package com.softwaremill.bootzooka.user
+/*
+ * COPYRIGHT (c) 2016 VOCADO, LLC.  ALL RIGHTS RESERVED.  THIS SOFTWARE CONTAINS
+ * TRADE SECRETS AND/OR CONFIDENTIAL INFORMATION PROPRIETARY TO VOCADO, LLC AND/OR
+ * ITS LICENSORS. ACCESS TO AND USE OF THIS INFORMATION IS STRICTLY LIMITED AND
+ * CONTROLLED BY VOCADO, LLC.  THIS SOFTWARE MAY NOT BE COPIED, MODIFIED, DISTRIBUTED,
+ * DISPLAYED, DISCLOSED OR USED IN ANY WAY NOT EXPRESSLY AUTHORIZED BY VOCADO, LLC IN WRITING.
+ */
 
-import com.softwaremill.bootzooka.user.domain.User
+package com.softwaremill.bootzooka.user.domain
+
 import org.scalatest.{FlatSpec, Matchers}
 
 class UserSpec extends FlatSpec with Matchers {
